Move Participation from models.people to models.participation
